pyproject.toml
[project]
name = "Hackathon"
version = "0.1.0"
description = "Hackathon backend"
requires-python = ">=3.12.6"
dependencies = [
    "fastapi[standard]==0.135.3",
    "sqlalchemy[asyncio]==2.0.49",
    "asyncpg==0.31.0",
    "python-dotenv==1.2.2",
    "uvicorn==0.44.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "bcrypt==5.0.0",
    "passlib==1.7.4",
    "authx==1.5.2",
    "python-jose[cryptography]"
]

[dependency-groups]
dev = [
    "black==26.3.1",
    "ruff==0.15.9",
    "mypy==1.20.0"
]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "W", "I", "C", "B", "A", "RUF"]
lint.ignore = ["E501"]
target-version = "py312"
exclude = ["tests/*", "pyproject.toml", ".venv/*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true 
disallow_untyped_defs = true
strict = true
